The browerWindow in saWin32Main.c is a pointer: it needs to be at least a long - which works for LP32 ILP32 LP64 and ILP64 Note: for LLP64 this should be long long, so atoll but we do not support it yet.

+++ platforms/win32/vm/sqWin32Main.c
@@ -1793,10 +1793,10 @@ parseVMArgument(int argc, char *argv[])
 
   /* NOTE: the following flags are "undocumented" */
 	else if (argc > 1 && !strcmp(argv[0], "-browserWindow")) {
-		browserWindow = (HWND)atoi(argv[1]);
+		browserWindow = (HWND)atol(argv[1]);
 		return 2; }
 	else if (!strncmp(argv[0], "-browserWindow:", 15)) {
-		browserWindow = (HWND)atoi(argv[0]+15);
+		browserWindow = (HWND)atol(argv[0]+15);
 		return 1; }
 
 	/* service support on 95 */
